groq-voice-assistant/requirements.txt
groq
sounddevice
soundfile
numpy
speechrecognition
pyttsx3
python-dotenv